BitStamp. Level1 refresh interval is 10 sec.
