+++ Connectors/BitStamp/BitStampMessageAdapter.cs
@@ -5,7 +5,6 @@ namespace StockSharp.BitStamp
 
 	using Ecng.Common;
 
-	using StockSharp.Algo;
 	using StockSharp.BitStamp.Native;
 	using StockSharp.Messages;
 	using StockSharp.Localization;
@@ -92,6 +91,7 @@ namespace StockSharp.BitStamp
 					_hasActiveOrders = false;
 					_hasMyTrades = false;
 					_requestOrderFirst = false;
+					_prevLevel1Time = default(DateTimeOffset);
 
 					if (_httpClient != null)
 					{
@@ -202,37 +202,7 @@ namespace StockSharp.BitStamp
 						ProcessPortfolioLookup(null);
 					}
 
-					var btcUsd = _httpClient.RequestBtcUsd();
-
-					if (btcUsd != null)
-					{
-						SendOutMessage(new Level1ChangeMessage
-						{
-							SecurityId = _btcUsd,
-							ServerTime = btcUsd.Time.ApplyTimeZone(TimeZoneInfo.Utc)
-						}
-						.TryAdd(Level1Fields.HighBidPrice, (decimal)btcUsd.High)
-						.TryAdd(Level1Fields.LowAskPrice, (decimal)btcUsd.Low)
-						.TryAdd(Level1Fields.VWAP, (decimal)btcUsd.VWAP)
-						.TryAdd(Level1Fields.LastTradePrice, (decimal)btcUsd.Last)
-						.TryAdd(Level1Fields.Volume, (decimal)btcUsd.Volume)
-						.TryAdd(Level1Fields.BestBidPrice, (decimal)btcUsd.Bid)
-						.TryAdd(Level1Fields.BestAskPrice, (decimal)btcUsd.Ask));
-					}
-
-					var eurUsd = _httpClient.RequestEurUsd();
-
-					if (eurUsd != null)
-					{
-						SendOutMessage(new Level1ChangeMessage
-						{
-							SecurityId = _eurUsd,
-							ServerTime = CurrentTime.Convert(TimeZoneInfo.Utc),
-						}
-						.TryAdd(Level1Fields.BestBidPrice, (decimal)eurUsd.Buy)
-						.TryAdd(Level1Fields.BestAskPrice, (decimal)eurUsd.Sell));
-					}
-
+					ProcessLevel1();
 					break;
 				}
